% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.r
\name{add_empty}
\alias{add_empty}
\title{Add Empty Cell}
\usage{
add_empty(list, item, index)
}
\arguments{
\item{list}{List to be used.}

\item{item}{String to be added as name of item in list.}

\item{index}{Integer. Index of item to be added.}
}
\value{
List. New list with item added with '-' as value.
}
\description{
Helper function to add an empty cell in a list
}
\examples{
l <- list(c('a', 'b', 'c'))
add_empty(l, 'd', 2)
}
