% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.r
\name{get_cluster_id}
\alias{get_cluster_id}
\title{Helper Get Cluster ID}
\usage{
get_cluster_id(num)
}
\arguments{
\item{num}{Integer or String. Can be either the number of the Cluster or
the name of cluster in format 'cluster-1'}
}
\value{
String of cluster-id in format 'cluster-1'
}
\description{
Helper function that gets format of Cluster ID to be accepted in VORTX.
}
\examples{
clusterid <- 'cluster-1'

get_cluster_id(clusterid)

get_cluster_id(1)
}
\keyword{internal}
