% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.r
\name{get_col}
\alias{get_col}
\title{Helper Get Column Name}
\usage{
get_col(col)
}
\arguments{
\item{col}{String or Vector of Strings. Cols to be formatted. Can be
the name of a column or a vector with names of columns.}
}
\value{
String in format '["col1", "col2", "col3"]'
}
\description{
Helper function that gets format of columns to be accepted in VORTX.
}
\examples{
cols <- c('col1', 'col2')
col <- 'col3'

get_col(cols)

get_col(col)
}
\keyword{internal}
