% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_create.r
\name{get_id_column}
\alias{get_id_column}
\title{Get ID Column}
\usage{
get_id_column(data, id = 1)
}
\arguments{
\item{data}{DataFrame to be checked}

\item{id}{Integer or String. This will be checked as possible ID. Default is 1.}
}
\value{
DataFrame with ID column first.
}
\description{
Defines which column will be used as ID in the dataset and puts it first named as 'id'
}
\examples{
\dontrun{
df <- r2vortx::wine
get_id_column(df)
}
}
\keyword{internal}
