% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_create.r
\name{get_ignored}
\alias{get_ignored}
\title{Get columns that should be ignored in VORTX}
\usage{
get_ignored(data)
}
\arguments{
\item{data}{data to be tested}
}
\value{
vector with ignored columns
}
\description{
Get columns that should be ignored in VORTX
}
