% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.r
\name{get_job_id}
\alias{get_job_id}
\title{Helper Get Job ID}
\usage{
get_job_id(job)
}
\arguments{
\item{job}{String List of parsed JSON. Can be either the job ID itself
in String format or a parsed JSON (result of a vortx_create_job, for example)
of a job.}
}
\value{
String of job ID
}
\description{
Helper function that gets format of job ID to be accepted in VORTX.
}
\examples{
jobid <- '1234567890'
job <- vortx_create_job(...)

get_job_id(jobid)

get_job_id(job)
}
