% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joblist.r
\name{get_joblist}
\alias{get_joblist}
\title{Raw list of jobs in VORTX}
\usage{
get_joblist(key, archived = "true", unarchived = "true")
}
\arguments{
\item{key}{String. User API Key for VORTX.}

\item{archived}{String Shows archived jobs. Default: 'true'}

\item{unarchived}{String Shows unarchived jobs. Default: 'true'}
}
\value{
DataFrame. Available jobs with their respective Job IDs, name and description.
}
\description{
Gets a raw list of jobs in VORTX user's account
}
\examples{
mykey <- '1234567890abcefghijkl'

get_joblist(mykey)
}
