% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.r
\name{get_metrics}
\alias{get_metrics}
\title{Get Metrics of Clusters}
\usage{
get_metrics(key, job, sandbox = FALSE)
}
\arguments{
\item{key}{String. User API Key for VORTX.}

\item{job}{String or List. Can be either a job ID number
in string format or parsed JSON in list format,
result of organizer or discoverer functions.}

\item{sandbox}{Choose TRUE if job is to be sent to sandbox instead}
}
\value{
DataFrame.
}
\description{
Gets metrics of clusters of job in VORTX in DataFrame format
}
\examples{
\dontrun{
mykey <- '1234567890abcefghijkl'
myjobid <- '0987654321'

myjobname <- 'My job'
myjobdesc <- 'This is a job that does job stuff'
df <- data.frame(c(1,2,3), c(4,5,6), c(7,8,9))
myjob <- vortx_create_job(mykey, df, myjobname, myjobdesc)

get_metrics(mykey, myjob)
}
}
\keyword{internal}
