% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.r
\name{get_metrics_raw}
\alias{get_metrics_raw}
\title{Get Raw Metrics of Clusters}
\usage{
get_metrics_raw(key, job)
}
\arguments{
\item{key}{String. User API Key for VORTX.}

\item{job}{String or List. Can be either a job ID number
in string format or parsed JSON in list format,
result of organizer or discoverer functions.}
}
\value{
List of parsed JSON.
}
\description{
Gets metrics of clusters of job in VORTX in parsed JSON format.
}
\examples{
mykey <- '1234567890abcefghijkl'
myjobid <- '0987654321'
myjob <- vortx_create_job(...)

get_metrics_raw(mykey, myjob)
}
