% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_create.r
\name{get_source}
\alias{get_source}
\title{Get file from different source if source is not NULL}
\usage{
get_source(data, source = NULL, sheet = NULL)
}
\arguments{
\item{data}{Depends on source. DataFrame for R, virtually anything that
identifies a googlesheet for 'googlesheets' or path for 'excel'.}

\item{source}{String defining source. May contain 'excel' or 'googlesheets'. Default NULL for R}

\item{sheet}{number or name of sheet to be imported from source. Default NULL for first.}
}
\value{
DataFrame to be sent to vortx
}
\description{
Get file from different source if source is not NULL
}
