% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_shiny.r
\name{get_summary}
\alias{get_summary}
\title{Get summary of numericals and categoricals}
\usage{
get_summary(df)
}
\arguments{
\item{df}{data.frame to be summarized}
}
\value{
list with summary for Numericals, Categories and individual factors for each category
}
\description{
Get summary of numericals and categoricals
}
\keyword{internal}
