% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.r
\name{get_summaryview_raw}
\alias{get_summaryview_raw}
\title{Get Raw Summaryview}
\usage{
get_summaryview_raw(key, job, clusternum, sandbox = FALSE)
}
\arguments{
\item{key}{String. User API Key for VORTX.}

\item{job}{String or List. Can be either a job ID number
in string format or parsed JSON in list format,
result of organizer or discoverer functions.}

\item{clusternum}{Integer or String. Can be cluster number (int) or cluster name in format 'cluster-x'.}

\item{sandbox}{Choose TRUE if job is to be sent to sandbox instead}
}
\value{
List of parsed JSON.
}
\description{
Gets raw summaryview of job in VORTX in parsed JSON format
}
\examples{
\dontrun{
mykey <- '1234567890abcefghijkl'
myjobid <- '0987654321'

myjobname <- 'My job'
myjobdesc <- 'This is a job that does job stuff'
df <- data.frame(c(1,2,3), c(4,5,6), c(7,8,9))
myjob <- vortx_create_job(mykey, df, myjobname, myjobdesc)

get_summaryview_raw(mykey, myjob, 1, 1, 10)
}
}
\keyword{internal}
