% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.r
\name{get_summaryview_single}
\alias{get_summaryview_single}
\title{Get Cluster's Summaryview}
\usage{
get_summaryview_single(key, job, clusternum)
}
\arguments{
\item{key}{String. User API Key for VORTX.}

\item{job}{String or List. Can be either a job ID number
in string format or parsed JSON in list format,
result of organizer or discoverer functions.}

\item{clusternum}{Integer or String. Can be cluster number (int) or cluster name in format 'cluster-x'.}
}
\value{
List of DataFrames.
}
\description{
Gets summaryview for a cluster of job in VORTX in list of DataFrame
}
\examples{
mykey <- '1234567890abcefghijkl'
myjobid <- '0987654321'
myjob <- vortx_create_job(...)

get_summaryview_single(mykey, myjob, 1)
}
