% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_create.r
\name{get_target_column}
\alias{get_target_column}
\title{Move Target column to first after ID}
\usage{
get_target_column(data, target)
}
\value{
DataFrame with target column after ID
}
\description{
Move Target column to first after ID
}
\keyword{internal}
