% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_create.r
\name{is_idish}
\alias{is_idish}
\title{Is IDish}
\usage{
is_idish(data, col)
}
\arguments{
\item{data}{DataFrame with column to be checked.}

\item{col}{Integer or String. Column to be checked.}
}
\value{
Boolean
}
\description{
Check if column has same number of unique elements as number of rows
}
