% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refine.r
\name{refine_clear_all}
\alias{refine_clear_all}
\title{Clears projects of determined name}
\usage{
refine_clear_all(project_name = "temp_project",
  base_url = "localhost:3333")
}
\arguments{
\item{project_name}{String wtih project name}

\item{base_url}{Base url for openrefine server}
}
\description{
Clears projects of determined name
}
\keyword{internal}
