% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refine.r
\name{refine_kill}
\alias{refine_kill}
\title{Delete OpenRefine project}
\usage{
refine_kill(name = NULL, id = NULL)
}
\arguments{
\item{name}{String. Name of project in OpenRefine.}

\item{id}{String. ID of project in OpenRefine. Obtainable by clicking 'About' in OpenRefine.}
}
\description{
Kills project in OpenRefine
}
