% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refine.r
\name{refine_mime}
\alias{refine_mime}
\title{Apply (mimes) changes from existent project in OpenRefine to R data}
\usage{
refine_mime(data, mime.name = NULL, mime.id = NULL)
}
\arguments{
\item{data}{DataFrame to be mimed}

\item{mime.name}{String with name of project that contains the changes. Either name or ID must be given.}

\item{mime.id}{String with ID of project that contains the changes. Either name or ID must be given.}
}
\description{
Apply (mimes) changes from existent project in OpenRefine to R data
}
\examples{
\dontrun{
refine_mime(df, 'diamonds')
}
}
