% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refine.r
\name{refine_pull}
\alias{refine_pull}
\title{Downloads csv file from OpenRefine}
\usage{
refine_pull(id = NULL, name = NULL, format = "csv", encoding = "UTF-8")
}
\arguments{
\item{id}{String. ID of project in OpenRefine. Obtainable by clicking 'About' in OpenRefine.}

\item{name}{String. Name of project in OpenRefine.}

\item{encoding}{String. Encoding of uploaded file. Default is UTF-8.}
}
\value{
DataFrame of dataset in OpenRefine
}
\description{
Extracts a dataset in OpenRefine to R as CSV and loaded as a DataFrame
}
\examples{
\dontrun{
df <- refine_pull('1622372310646')
}
}
