% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_create.r
\name{remove_constant}
\alias{remove_constant}
\title{Delete column with only one value}
\usage{
remove_constant(data)
}
\value{
List with DataFrame with removed column and the name of the column
}
\description{
Delete column with only one value
}
\keyword{internal}
