% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_create.r
\name{rename_ignored}
\alias{rename_ignored}
\title{Ignore variables}
\usage{
rename_ignored(data, ignoredcols)
}
\arguments{
\item{data}{data with variables to ignore}

\item{ignoredcols}{character vector with names of variables to be ignored}
}
\value{
new data with renamed variables
}
\description{
Ignore variables
}
\keyword{internal}
