% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizer.r
\name{start_organizer}
\alias{start_organizer}
\title{Start Organizer}
\usage{
start_organizer(key, job, ignoredcols = NULL)
}
\arguments{
\item{key}{String. User API Key for VORTX.}

\item{job}{String or List. Can be either a job ID number
in string format or parsed JSON in list format,
result of organizer or discoverer functions.}

\item{ignoredcols}{String or Vector of strings. Optional name of columns to be ignored. Default is NULL.}
}
\value{
Job. Parsed content of API request, containing job information, such as job ID, used in other functions.
}
\description{
Runs Organizer from a created Job in VORTX
}
\examples{
\dontrun{
mykey <- '1234567890abcefghijkl'
myjobid <- '0987654321'

myjobname <- 'My job'
myjobdesc <- 'This is a job that does job stuff'
df <- data.frame(c(1,2,3), c(4,5,6), c(7,8,9))
myjob <- create_job(mykey, df, myjobname, myjobdesc)

start_organizer(mykey, myjob, 'WhateverCol')
}
}
