% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.r
\name{vortx_dataset}
\alias{vortx_dataset}
\title{Get full dataset}
\usage{
vortx_dataset(key, job)
}
\arguments{
\item{key}{String. User API Key for VORTX.}

\item{job}{String or List. Can be either a job ID number
in string format or parsed JSON in list format,
result of organizer or discoverer functions.}
}
\value{
DataFrame of a job dataset
}
\description{
Gets full dataset of job in VORTX in DataFrame format
}
\examples{
mykey <- '1234567890abcefghijkl'
myjobid <- '0987654321'
myjob <- vortx_create_job(...)

vortx_dataset(mykey, myjob)
}
