% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discoverer.r
\name{vortx_discoverer}
\alias{vortx_discoverer}
\title{Run Discoverer from Scratch}
\usage{
vortx_discoverer(key, data, jobname, target, jobdesc = NULL,
  ignoredcols = NULL)
}
\arguments{
\item{key}{String. User API Key for VORTX.}

\item{data}{DataFrame. Data to be created as a job.}

\item{jobname}{String. Title of job to be created.}

\item{target}{String. Name of column to be used as Target.}

\item{jobdesc}{String. Description of job to be created. Optional. Default NULL.}

\item{ignoredcols}{String or Vector of strings. Optional name of columns to be ignored. Default is NULL.}
}
\value{
Job. Parsed content of API request, containing job information, such as job ID, used in other functions.
}
\description{
Creates a job and runs Discoverer
}
\examples{
mykey <- '1234567890abcefghijkl'
myjobname <- 'My job'
myjobdesc <- 'This is a job that does job stuff'
df <- r2vortx::wine

vortx_discoverer(mykey, df, myjobname, 'Alcohol', myjobdesc, 'Ash')
}
