% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobinfo.r
\name{vortx_info}
\alias{vortx_info}
\title{Get information}
\usage{
vortx_info(key, job, info = "all")
}
\arguments{
\item{key}{String. User API Key for VORTX.}

\item{job}{String or List. Can be either a job ID number
in string format or parsed JSON in list format,
result of organizer or discoverer functions.}

\item{info}{String. Defines what kind of information to be extracted from job,
between 'clusters', 'varscores', 'dataset' or 'summary'. Default is 'all'.}
}
\value{
List of DataFrames or DataFrame. If 'all' or 'summary' are defined, a list. Else specific DataFrame.
}
\description{
Gets information on a VORTX job
}
\examples{
mykey <- '1234567890abcefghijkl'
myjobid <- '0987654321'
myjob <- vortx_create_job(...)

vortx_info(mykey, myjobid, info = 'clusters')
vortx_info(mykey, myjob)
}
