% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizer.r
\name{vortx_organizer}
\alias{vortx_organizer}
\title{Run Organizer from Scratch}
\usage{
vortx_organizer(key, data, jobname, jobdesc = NULL, ignoredcols = NULL,
  id = 1)
}
\arguments{
\item{key}{String. User API Key for VORTX.}

\item{data}{DataFrame. Data to be created as a job.}

\item{jobname}{String. Title of job to be created.}

\item{jobdesc}{String. Description of job to be created. Optional. Default NULL.}

\item{ignoredcols}{String or Vector of strings. Optional name of columns to be ignored. Default is NULL.}

\item{id}{Integer or String. This will be checked as possible ID. Default is 1.}
}
\value{
Job. Parsed content of API request, containing job information, such as job ID, used in other functions.
}
\description{
Creates a job and runs Organizer
}
\examples{
\dontrun{
mykey <- '1234567890abcefghijkl'
myjobname <- 'My job'
myjobdesc <- 'This is a job that does job stuff'
df <- r2vortx::wine

vortx_organizer(mykey, df, myjobname, myjobdesc, 'Ash')
}
}
