% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.r
\docType{data}
\name{wine}
\alias{wine}
\title{Dataset of Wine}
\format{A data frame with 178 rows and 14 variables.}
\usage{
wine
}
\description{
A dataset containing characteristics of wines.
}
\keyword{datasets}
