% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createjob.r
\name{create_job}
\alias{create_job}
\title{Create job}
\usage{
create_job(key, data, jobname, jobdesc = NULL,
  vortx_server = "production")
}
\arguments{
\item{key}{String. User API Key for VORTX.}

\item{data}{DataFrame. Data to be created as a job.}

\item{jobname}{String. Title of job to be created.}

\item{jobdesc}{String. Description of job to be created. Optional. Default NULL.}

\item{vortx_server}{Choose server to run Vortx. Can be one of "production", "sandbox", "local" or desired URL.}
}
\value{
Job. Parsed content of API request, containing job information, such as job ID, used in other functions.
}
\description{
Creates a job in VORTX
}
\examples{
\dontrun{
mykey <- '1234567890abcefghijkl'
myjobname <- 'My job'
myjobdesc <- 'This is a job that does job stuff'
df <- r2vortx::wine

create_job(mykey, df, myjobname, myjobdesc)
}
}
