% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.r
\name{get_dataset_single}
\alias{get_dataset_single}
\title{Get dataset for a single cluster}
\usage{
get_dataset_single(key, job, clusternum, start, end, sandbox = FALSE)
}
\arguments{
\item{key}{String. User API Key for VORTX.}

\item{job}{String or List. Can be either a job ID number
in string format or parsed JSON in list format,
result of organizer or discoverer functions.}

\item{clusternum}{Integer or String. Can be cluster number (int) or cluster name in format 'cluster-x'.}

\item{start}{Integer. Starting row to get info from.}

\item{end}{Integer. Ending row to get info from.}

\item{sandbox}{Choose TRUE if job is to be sent to sandbox instead}
}
\value{
DataFrame of a cluster dataset
}
\description{
Gets dataset for a cluster of job in VORTX in DataFrame format
}
\examples{
\dontrun{
mykey <- '1234567890abcefghijkl'
myjobid <- '0987654321'

myjobname <- 'My job'
myjobdesc <- 'This is a job that does job stuff'
df <- data.frame(c(1,2,3), c(4,5,6), c(7,8,9))
myjob <- vortx_create_job(mykey, df, myjobname, myjobdesc)

get_dataset_single(mykey, myjob, 1, 1, 10)
}
}
