% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchy.r
\name{get_hierarchy}
\alias{get_hierarchy}
\title{Get Hierarchy of Clusters}
\usage{
get_hierarchy(key, job)
}
\arguments{
\item{key}{String. User API Key for VORTX.}

\item{job}{String or List. Can be either a job ID number 
in string format or parsed JSON in list format, 
result of organizer or discoverer functions.}
}
\value{
DataFrame.
}
\description{
Gets hierarchy of clusters of job in VORTX in DataFrame format
}
\examples{
mykey <- '1234567890abcefghijkl'
myjobid <- '0987654321'
myjob <- vortx_create_job(...)

> get_hierarchy(mykey, myjob)
}
