% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refine.r
\name{refine_list}
\alias{refine_list}
\title{Gets list of projects in OpenRefine}
\usage{
refine_list(base_url = "http://localhost:3333")
}
\arguments{
\item{base_url}{Base url for openrefine server}
}
\value{
DataFrame with projects' ID, Name, Last Modified and Number of Rows
}
\description{
Gets list of projects in OpenRefine
}
