% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refine.r
\name{refine_pull}
\alias{refine_pull}
\title{Pull data from OpenRefine to R}
\usage{
refine_pull(name = NULL, id = NULL, encoding = "UTF-8")
}
\arguments{
\item{name}{String. Name of project in OpenRefine. Either name or ID must be given.}

\item{id}{String. ID of project in OpenRefine. Obtainable by clicking 'About' in OpenRefine. Either name or ID must be given.}

\item{encoding}{String. Encoding of uploaded file. Default is UTF-8.}
}
\value{
DataFrame of dataset in OpenRefine
}
\description{
Pulls data from OpenRefine to R as a DataFrame
}
\examples{
\dontrun{
df <- refine_pull('1622372310646')
}
}
