% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discoverer.r
\name{start_discoverer}
\alias{start_discoverer}
\title{Start Discoverer}
\usage{
start_discoverer(key, job, target, ignoredcols = NULL)
}
\arguments{
\item{key}{String. User API Key for VORTX.}

\item{job}{String or List. Can be either a job ID number 
in string format or parsed JSON in list format, 
result of organizer or discoverer functions.}

\item{target}{String. Name of column to be used as Target.}

\item{ignoredcols}{String or Vector of strings. Optional name of columns to be ignored. Default is NULL.}
}
\value{
Job. Parsed content of API request, containing job information, such as job ID, used in other functions.
}
\description{
Runs Discoverer from a created Job in VORTX
}
\examples{
mykey <- '1234567890abcefghijkl'
myjobid <- '0987654321'
myjob <- vortx_create_job(...)

> start_discoverer(mykey, myjob, myjobname, myjobdesc)
}
