% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizer.r
\name{start_organizer}
\alias{start_organizer}
\title{Start Organizer}
\usage{
start_organizer(key, job, ignoredcols = NULL)
}
\arguments{
\item{key}{String. User API Key for VORTX.}

\item{job}{String or List. Can be either a job ID number
in string format or parsed JSON in list format,
result of organizer or discoverer functions.}

\item{ignoredcols}{String or Vector of strings. Optional name of columns to be ignored. Default is NULL.}
}
\value{
Job. Parsed content of API request, containing job information, such as job ID, used in other functions.
}
\description{
Runs Organizer from a created Job in VORTX
}
\examples{
mykey <- '1234567890abcefghijkl'
myjobid <- '0987654321'
myjob <- vortx_create_job(...)

start_organizer(mykey, myjob, 'WhateverCol')
}
