% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discoverer.r
\name{vortx_discoverer}
\alias{vortx_discoverer}
\title{Run Discoverer from Scratch}
\usage{
vortx_discoverer(key, data, jobname, target, jobdesc = NULL,
  ignoredcols = NULL, id = 1, source = "r", sheet = NULL,
  sandbox = FALSE)
}
\arguments{
\item{key}{String. User API Key for VORTX.}

\item{data}{Data to be created as a job. If no source is defined, it should be a DataFrame.
If source is 'googlesheets', could be name, ID or any information that refers to it.
If source is 'excel', it should be the path to file.}

\item{jobname}{String. Title of job to be created.}

\item{target}{String. Name of column to be used as Target.}

\item{jobdesc}{String. Description of job to be created. Optional. Default NULL.}

\item{ignoredcols}{String or Vector of strings. Optional name of columns to be ignored. Default is NULL.}

\item{id}{Integer or String. This will be checked as possible ID. Default is 1.}

\item{source}{String defining source. May contain 'excel' or 'googlesheets'. Default NULL for R.
Use 'googlesheets_new' for new user.}

\item{sheet}{String with number or name of sheet to be imported from source. Default NULL for first. Unnecessary for R.}

\item{sandbox}{Choose TRUE if job is to be sent to sandbox instead}
}
\value{
Job. Parsed content of API request, containing job information, such as job ID, used in other functions.
}
\description{
Creates a job and runs Discoverer
}
\examples{
\dontrun{
mykey <- '1234567890abcefghijkl'
myjobname <- 'My job'
myjobdesc <- 'This is a job that does job stuff'
df <- r2vortx::wine

vortx_discoverer(mykey, df, myjobname, 'Alcohol', myjobdesc, 'Ash')
}
}
