% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobinfo.r
\name{vortx_info}
\alias{vortx_info}
\title{Get information}
\usage{
vortx_info(key, job, info = "all", vortx_server = "production")
}
\arguments{
\item{key}{String. User API Key for VORTX.}

\item{job}{String or List. Can be either a job ID number
in string format or parsed JSON in list format,
result of organizer or discoverer functions.}

\item{info}{String. Defines what kind of information to be extracted from job,
between 'clusters', 'varscores', 'dataset' or 'summary'. Default is 'all'.}

\item{vortx_server}{Choose server to run Vortx. Can be one of "production", "sandbox", "local" or desired URL.}
}
\value{
List of DataFrames or DataFrame. If 'all' or 'summary' are defined, a list. Else specific DataFrame.
}
\description{
Gets information on a VORTX job
}
\examples{
\dontrun{
mykey <- '1234567890abcefghijkl'
myjobid <- '0987654321'

myjobname <- 'My job'
myjobdesc <- 'This is a job that does job stuff'
df <- data.frame(c(1,2,3), c(4,5,6), c(7,8,9))
myjob <- vortx_create_job(mykey, df, myjobname, myjobdesc)

vortx_info(mykey, myjobid, info = 'clusters')
vortx_info(mykey, myjob)
}
}
